backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.1
google-auth==2.29.0
google-cloud-vision==3.4.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.3.0
pyzbar==0.1.9
requests==2.32.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.30
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
